Address code review feedback: document deque rationale and fix size_t consistency
Agent-Logs-Url: https://github.com/Jonathhhan/ofxStableDiffusion/sessions/84a4d7bc-a610-4314-acdc-f77a3296554e

+++ src/core/ofxStableDiffusionProgressTracker.cpp
@@ -33,7 +33,7 @@ void ofxStableDiffusionProgressTracker::update(int currentStep, int currentBatch
 		stepTimes.push_back(stepTime);
 
 		// Keep only recent history
-		if (static_cast<int>(stepTimes.size()) > maxHistorySize) {
+		if (stepTimes.size() > maxHistorySize) {
 			stepTimes.pop_front();
 		}
 	}
